# Seal types not suitable for extending