--- a/Fluid/Accessors/DelegateAccessor.cs
+++ b/Fluid/Accessors/DelegateAccessor.cs
@@ -2,7 +2,7 @@
 
 namespace Fluid.Accessors
 {
-    public class DelegateAccessor : DelegateAccessor<object, object>
+    public sealed class DelegateAccessor : DelegateAccessor<object, object>
     {
         public DelegateAccessor(Func<object, string, object> getter) : base((obj, name, ctx) => getter(obj, name))
         {