--- a/Fluid/Ast/ContinueStatement.cs
+++ b/Fluid/Ast/ContinueStatement.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 
 namespace Fluid.Ast
 {
-    public class ContinueStatement : Statement
+    public sealed class ContinueStatement : Statement
     {
         public override ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
         {