--- a/Fluid/Ast/FunctionCallSegment.cs
+++ b/Fluid/Ast/FunctionCallSegment.cs
@@ -5,7 +5,7 @@ using Fluid.Values;
 
 namespace Fluid.Ast
 {
-    public class FunctionCallSegment : MemberSegment
+    public sealed class FunctionCallSegment : MemberSegment
     {
         private static readonly FunctionArguments NonCacheableArguments = new();
         private volatile FunctionArguments _cachedArguments;