--- a/Fluid/Ast/IndexerSegment.cs
+++ b/Fluid/Ast/IndexerSegment.cs
@@ -3,7 +3,7 @@ using Fluid.Values;
 
 namespace Fluid.Ast
 {
-    public class IndexerSegment : MemberSegment
+    public sealed class IndexerSegment : MemberSegment
     {
         public IndexerSegment(Expression expression)
         {