--- a/Fluid/Ast/LiquidStatement.cs
+++ b/Fluid/Ast/LiquidStatement.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 namespace Fluid.Ast
 {
-    public class LiquidStatement : TagStatement
+    public sealed class LiquidStatement : TagStatement
     {
         public LiquidStatement(List<Statement> statements) : base(statements)
         {