--- a/Fluid/Ast/NamedExpressionList.cs
+++ b/Fluid/Ast/NamedExpressionList.cs
@@ -2,7 +2,7 @@
 
 namespace Fluid.Ast
 {
-    public class NamedExpressionList
+    public sealed class NamedExpressionList
     {
         public static readonly NamedExpressionList Empty = new NamedExpressionList();