--- a/Fluid/Ast/OutputStatement.cs
+++ b/Fluid/Ast/OutputStatement.cs
@@ -6,7 +6,7 @@ using Fluid.Values;
 
 namespace Fluid.Ast
 {
-    public class OutputStatement : Statement
+    public sealed class OutputStatement : Statement
     {
         public OutputStatement(Expression expression)
         {