--- a/Fluid/ParseException.cs
+++ b/Fluid/ParseException.cs
@@ -2,7 +2,7 @@
 
 namespace Fluid
 {
-    public class ParseException : Exception
+    public sealed class ParseException : Exception
     {
         //
         // Summary: