@sebastienros 2023-12-08
I fear that it will break someone. But let's try and see people complain, the current version being quite stable they shouldn't need to upgrade and won't be blocked.