@lahma 2023-12-08
This was my thinking too, let's break it and unseal when needed.